Conflicts:
	.gitignore
	CycleStreets.xcodeproj/project.pbxproj
	CycleStreets.xcodeproj/project.xcworkspace/xcuserdata/neil.xcuserdatad/UserInterfaceState.xcuserstate
	CycleStreets.xcodeproj/xcuserdata/neil.xcuserdatad/xcdebugger/Breakpoints.xcbkptlist

+++ lib/categories/NSDate-Misc.h
@@ -15,7 +15,7 @@
 - (NSDate *)dateAsDateWithoutTime;
 - (NSDate *)dateWithZeroSeconds;
 - (NSDate *)dateWithZeroTime;
-- (int)differenceInDaysTo:(NSDate *)toDate;
+- (NSInteger)differenceInDaysTo:(NSDate *)toDate;
 - (NSString *)formattedDateString;
 - (NSString *)formattedStringUsingFormat:(NSString *)dateFormat;
 - (NSString *)YTTDFrom:(NSDate *)comparisonDate withFormat:(NSString *)dateFormat;
